Struct mpris_server::builder::MetadataBuilder

pub struct MetadataBuilder { /* private fields */ }

A builder used to create Metadata.

Implementations

impl MetadataBuilder

pub fn other(self, key: &str, value: impl Into<Value<'static>>) -> Self

Sets a value for the given key.

pub fn trackid(self, trackid: impl Into<TrackId>) -> Self

Sets a unique identity for this track within the context of an MPRIS object (eg: tracklist).

This contains a D-Bus path that uniquely identifies the track within the scope of the playlist. There may or may not be an actual D-Bus object at that path; this specification says nothing about what interfaces such an object may implement.

pub fn length(self, length: Time) -> Self

Sets the duration of the track.

pub fn art_url(self, art_url: impl Into<Uri>) -> Self

Sets the location of an image representing the track or album.

Clients should not assume this will continue to exist when the media player stops giving out the URL.

pub fn album(self, album: impl Into<String>) -> Self

Sets the album name.

pub fn album_artist( self, album_artist: impl IntoIterator<Item = impl Into<String>> ) -> Self

Sets the album artist(s).

pub fn artist(self, artist: impl IntoIterator<Item = impl Into<String>>) -> Self

Sets the track artist(s).

pub fn lyrics(self, lyrics: impl Into<String>) -> Self

Sets the track lyrics.

pub fn audio_bpm(self, audio_bpm: i32) -> Self

Sets the speed of the music, in beats per minute.

pub fn auto_rating(self, auto_rating: f64) -> Self

Sets an automatically-generated rating, based on things such as how often it has been played. This should be in the range 0.0 to 1.0.

pub fn comment( self, comment: impl IntoIterator<Item = impl Into<String>> ) -> Self

Sets a (list of) freeform comment(s).

pub fn composer( self, composer: impl IntoIterator<Item = impl Into<String>> ) -> Self

Sets the composer(s) of the track.

pub fn content_created(self, content_created: impl Into<DateTime>) -> Self

Sets when the track was created. Usually only the year component will be useful.

pub fn disc_number(self, disc_number: i32) -> Self

Sets the disc number on the album that this track is from.

pub fn first_used(self, first_used: impl Into<DateTime>) -> Self

Sets when the track was first played.

pub fn genre(self, genre: impl IntoIterator<Item = impl Into<String>>) -> Self

Sets the genre(s) of the track.

pub fn last_used(self, last_used: impl Into<DateTime>) -> Self

Sets when the track was last played.

pub fn lyricist( self, lyricist: impl IntoIterator<Item = impl Into<String>> ) -> Self

Sets the lyricist(s) of the track.

pub fn title(self, title: impl Into<String>) -> Self

Sets the track title.

pub fn track_number(self, track_number: i32) -> Self

Sets the track number on the album disc.

pub fn url(self, url: impl Into<Uri>) -> Self

Sets the location of the media file.

pub fn use_count(self, use_count: i32) -> Self

Sets the number of times the track has been played.

pub fn user_rating(self, user_rating: f64) -> Self

Sets a user-specified rating. This should be in the range 0.0 to 1.0.

pub fn build(self) -> Metadata

Creates Metadata from the builder.