mpris_server

Struct Time

pub struct Time(/* private fields */);

A time with microsecond precision which can be negative.

Implementations

impl Time

pub const ZERO: Self

A time of zero.

Examples

use mpris_server::Time;

let time = Time::ZERO;
assert!(time.is_zero());
assert_eq!(time.as_micros(), 0);

pub const MIN: Self

The minimum time.

pub const MAX: Self

The maximum time.

pub const fn from_secs(secs: i64) -> Self

Creates a new Time from the specified number of whole seconds.

Panics

This will panic if the number of seconds overflows.

Examples

use mpris_server::Time;

assert_eq!(Time::from_secs(5).as_nanos(), 5_000_000_000);

pub const fn from_millis(millis: i64) -> Self

Creates a new Time from the specified number of whole milliseconds.

Panics

This will panic if the number of milliseconds overflows.

Examples

use mpris_server::Time;

assert_eq!(Time::from_millis(5).as_nanos(), 5_000_000);

pub const fn from_micros(micros: i64) -> Self

Creates a new Time from the specified number of whole microseconds.

Examples

use mpris_server::Time;

assert_eq!(Time::from_micros(5).as_nanos(), 5_000);

pub const fn from_nanos(nanos: i64) -> Self

Creates a new Time from the specified number of whole nanoseconds.

Note: This will round of the nanoseconds to microseconds level of precision.

Examples

use mpris_server::Time;

assert_eq!(Time::from_nanos(5).as_nanos(), 0);
assert_eq!(Time::from_nanos(5342).as_nanos(), 5000);

pub const fn as_secs(&self) -> i64

Returns the number of whole seconds contained by this Time.

Examples

use mpris_server::Time;

assert_eq!(Time::from_micros(5_000_000).as_secs(), 5);
assert_eq!(Time::from_micros(3).as_secs(), 0);

pub const fn as_millis(&self) -> i64

Returns the number of whole milliseconds contained by this Time.

Examples

use mpris_server::Time;

assert_eq!(Time::from_micros(5_000_000).as_millis(), 5_000);
assert_eq!(Time::from_micros(3).as_millis(), 0);

pub const fn as_micros(&self) -> i64

Returns the number of whole microseconds contained by this Time.

Examples

use mpris_server::Time;

assert_eq!(Time::from_micros(5_000_000).as_micros(), 5_000_000);
assert_eq!(Time::from_micros(3).as_micros(), 3);

pub const fn as_nanos(&self) -> i128

Returns the number of whole nanoseconds contained by this Time.

Examples

use mpris_server::Time;

assert_eq!(Time::from_micros(5_000_000).as_nanos(), 5_000_000_000);
assert_eq!(Time::from_micros(3).as_nanos(), 3_000);

pub const fn is_zero(&self) -> bool

Returns true if this Time is zero.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_zero(), true);
assert_eq!(Time::from_micros(1).is_zero(), false);

pub const fn is_negative(&self) -> bool

Returns true if this Time is negative.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_negative(), false);
assert_eq!(Time::from_micros(-1).is_negative(), true);
assert_eq!(Time::from_micros(1).is_negative(), false);

pub const fn is_positive(&self) -> bool

Returns true if this Time is positive.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_positive(), false);
assert_eq!(Time::from_micros(1).is_positive(), true);
assert_eq!(Time::from_micros(-1).is_positive(), false);

pub const fn abs(&self) -> Self

Returns the time as an absolute (non-negative) value.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.abs(), Time::ZERO);
assert_eq!(Time::from_micros(-1).abs(), Time::from_micros(1));
assert_eq!(Time::from_micros(1).abs(), Time::from_micros(1));

pub const fn checked_add(self, other: Self) -> Option<Self>

Checked Time addition. Computes self + other, returning None if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(1).checked_add(Time::from_micros(1)),
    Some(Time::from_micros(2))
);
assert_eq!(Time::MAX.checked_add(Time::from_micros(1)), None);

pub const fn checked_sub(self, other: Self) -> Option<Self>

Checked Time subtraction. Computes self - other, returning None if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(2).checked_sub(Time::from_micros(1)),
    Some(Time::from_micros(1))
);
assert_eq!(Time::MIN.checked_sub(Time::from_micros(1)), None);

pub const fn saturating_add(self, other: Self) -> Self

Saturating Time addition. Computes self + other, returning Time::MAX if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(1).saturating_add(Time::from_micros(1)),
    Time::from_micros(2)
);
assert_eq!(Time::MAX.saturating_add(Time::from_micros(1)), Time::MAX);

pub const fn saturating_sub(self, other: Self) -> Self

Saturating Time subtraction. Computes self - other, returning Time::MIN if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(2).saturating_sub(Time::from_micros(1)),
    Time::from_micros(1)
);
assert_eq!(Time::MIN.saturating_sub(Time::from_micros(1)), Time::MIN);