Struct mpris_server::Metadata

pub struct Metadata(_);

A mapping from metadata attribute names to values.

The mpris:trackid attribute must always be present.

If the length of the track is known, it should be provided in the metadata property with the mpris:length key.

If there is an image associated with the track, a URL for it may be provided using the mpris:artUrl key.

Implementations

impl Metadata

pub fn new() -> Self

Create an empty Metadata.

pub fn builder() -> MetadataBuilder

Creates a new builder-pattern struct instance to construct Metadata.

pub fn insert( &mut self, key: impl Into<String>, value: impl Into<Value<'static>> ) -> Option<Value<'static>>

Insert a new key-value pair into the metadata.

This will overwrite any existing value for the given key and will return the old value, if any.

pub fn get<'v, V>(&'v self, key: &str) -> Option<Result<&'v V>>where &'v V: TryFrom<&'v Value<'v>>,

Get the value of the given key and convert it to V, if it exists.

pub fn get_value(&self, key: &str) -> Option<&Value<'_>>

Get the value of the given key, if it exists.

pub fn set_trackid(&mut self, trackid: impl Into<OwnedObjectPath>)

A unique identity for this track within the context of an MPRIS object (eg: tracklist).

This contains a D-Bus path that uniquely identifies the track within the scope of the playlist. There may or may not be an actual D-Bus object at that path; this specification says nothing about what interfaces such an object may implement.

pub fn set_length(&mut self, length: TimeInUs)

The duration of the track in microseconds.

pub fn set_art_url(&mut self, art_url: impl Into<Uri>)

The location of an image representing the track or album.

Clients should not assume this will continue to exist when the media player stops giving out the URL.

pub fn set_album(&mut self, album: impl Into<String>)

The album name.

pub fn set_album_artist( &mut self, album_artist: impl IntoIterator<Item = impl Into<String>> )

The album artist(s).

pub fn set_artist( &mut self, artist: impl IntoIterator<Item = impl Into<String>> )

The track artist(s).

pub fn set_lyrics(&mut self, lyrics: impl Into<String>)

The track lyrics.

pub fn set_audio_bpm(&mut self, audio_bpm: i32)

The speed of the music, in beats per minute.

pub fn set_auto_rating(&mut self, auto_rating: f64)

An automatically-generated rating, based on things such as how often it has been played. This should be in the range 0.0 to 1.0.

pub fn set_comment( &mut self, comment: impl IntoIterator<Item = impl Into<String>> )

A (list of) freeform comment(s).

pub fn set_composer( &mut self, composer: impl IntoIterator<Item = impl Into<String>> )

The composer(s) of the track.

pub fn set_content_created(&mut self, content_created: impl Into<DateTime>)

When the track was created. Usually only the year component will be useful.

pub fn set_disc_number(&mut self, disc_number: i32)

The disc number on the album that this track is from.

pub fn set_first_used(&mut self, first_used: impl Into<DateTime>)

When the track was first played.

pub fn set_genre(&mut self, genre: impl IntoIterator<Item = impl Into<String>>)

The genre(s) of the track.

pub fn set_last_used(&mut self, last_used: impl Into<DateTime>)

When the track was last played.

pub fn set_lyricist( &mut self, lyricist: impl IntoIterator<Item = impl Into<String>> )

The lyricist(s) of the track.

pub fn set_title(&mut self, title: impl Into<String>)

The track title.

pub fn set_track_number(&mut self, track_number: i32)

The track number on the album disc.

pub fn set_url(&mut self, url: impl Into<Uri>)

The location of the media file.

pub fn set_use_count(&mut self, use_count: i32)

The number of times the track has been played.

pub fn set_user_rating(&mut self, user_rating: f64)

A user-specified rating. This should be in the range 0.0 to 1.0.