Struct mpris_server::Time

pub struct Time(_);

A time which can be negative.

Implementations

impl Time

pub const ZERO: Self = _

A time of zero.

Examples

use mpris_server::Time;

let time = Time::ZERO;
assert!(time.is_zero());
assert_eq!(time.as_micros(), 0);

pub const MIN: Self = _

The minimum time.

pub const MAX: Self = _

The maximum time.

pub const fn from_secs(secs: i64) -> Self

Creates a new Time from the specified number of whole seconds.

pub const fn from_millis(millis: i64) -> Self

Creates a new Time from the specified number of whole milliseconds.

pub const fn from_micros(micros: i64) -> Self

Creates a new Time from the specified number of whole microseconds.

pub const fn from_nanos(nanos: i64) -> Self

Creates a new Time from the specified number of whole nanoseconds.

pub const fn as_secs(&self) -> i64

Returns the number of whole seconds contained by this Time.

pub const fn as_millis(&self) -> i64

Returns the number of whole milliseconds contained by this Time.

pub const fn as_micros(&self) -> i64

Returns the number of whole microseconds contained by this Time.

pub const fn as_nanos(&self) -> i64

Returns the number of whole nanoseconds contained by this Time.

pub const fn is_zero(&self) -> bool

Returns true if this Time is zero.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_zero(), true);
assert_eq!(Time::from_micros(1).is_zero(), false);

pub const fn is_negative(&self) -> bool

Returns true if this Time is negative.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_negative(), false);
assert_eq!(Time::from_micros(-1).is_negative(), true);
assert_eq!(Time::from_micros(1).is_negative(), false);

pub const fn is_positive(&self) -> bool

Returns true if this Time is positive.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.is_positive(), false);
assert_eq!(Time::from_micros(1).is_positive(), true);
assert_eq!(Time::from_micros(-1).is_positive(), false);

pub const fn abs(&self) -> Self

Returns the time as an absolute (non-negative) value.

Examples

use mpris_server::Time;

assert_eq!(Time::ZERO.abs(), Time::ZERO);
assert_eq!(Time::from_micros(-1).abs(), Time::from_micros(1));
assert_eq!(Time::from_micros(1).abs(), Time::from_micros(1));

pub const fn checked_add(self, other: Self) -> Option<Self>

Checked Time addition. Computes self + other, returning None if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(1).checked_add(Time::from_micros(1)),
    Some(Time::from_micros(2))
);
assert_eq!(Time::MAX.checked_add(Time::from_micros(1)), None);

pub const fn checked_sub(self, other: Self) -> Option<Self>

Checked Time subtraction. Computes self - other, returning None if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(2).checked_sub(Time::from_micros(1)),
    Some(Time::from_micros(1))
);
assert_eq!(Time::MIN.checked_sub(Time::from_micros(1)), None);

pub const fn saturating_add(self, other: Self) -> Self

Saturating Time addition. Computes self + other, returning Time::MAX if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(1).saturating_add(Time::from_micros(1)),
    Time::from_micros(2)
);
assert_eq!(Time::MAX.saturating_add(Time::from_micros(1)), Time::MAX);

pub const fn saturating_sub(self, other: Self) -> Self

Saturating Time subtraction. Computes self - other, returning Time::MIN if overflow occurred.

Examples

use mpris_server::Time;

assert_eq!(
    Time::from_micros(2).saturating_sub(Time::from_micros(1)),
    Time::from_micros(1)
);
assert_eq!(Time::MIN.saturating_sub(Time::from_micros(1)), Time::MIN);